Organized Persisters into proper namespaces. Added mapping support to association persisters. Prepared field for Custom Persisters. Initial draft of Entity and Collection Persister interfaces.
