+++ lib/Doctrine/ORM/Persister/Collection/AbstractCollectionPersister.php
@@ -17,7 +17,7 @@
  * <http://www.doctrine-project.org>.
  */
 
-namespace Doctrine\ORM\Persisters;
+namespace Doctrine\ORM\Persister\Collection;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\PersistentCollection;
@@ -51,7 +51,7 @@ abstract class AbstractCollectionPersister
      * @var \Doctrine\DBAL\Platforms\AbstractPlatform
      */
     protected $platform;
-    
+
     /**
      * The quote strategy.
      *
@@ -170,7 +170,7 @@ abstract class AbstractCollectionPersister
      * Counts the size of this persistent collection.
      *
      * @param \Doctrine\ORM\PersistentCollection $coll
-     * 
+     *
      * @return integer
      *
      * @throws \BadMethodCallException
@@ -258,10 +258,10 @@ abstract class AbstractCollectionPersister
 
     /**
      * Gets an element by key.
-     * 
+     *
      * @param \Doctrine\ORM\PersistentCollection $coll
      * @param mixed                              $index
-     * 
+     *
      * @return mixed
      *
      * @throws \BadMethodCallException
