+++ tests/Doctrine/Tests/Mocks/UnitOfWorkMock.php
@@ -42,8 +42,8 @@ class UnitOfWorkMock extends \Doctrine\ORM\UnitOfWork
      * Sets a (mock) persister for an entity class that will be returned when
      * getEntityPersister() is invoked for that class.
      *
-     * @param string                                        $entityName
-     * @param \Doctrine\ORM\Persisters\BasicEntityPersister $persister
+     * @param string                                              $entityName
+     * @param \Doctrine\ORM\Persister\Entity\BasicEntityPersister $persister
      *
      * @return void
      */
